Saving progress to merge and split the testing work
